scholarai/requirements.txt
streamlit>=1.32.0
openai>=1.30.0
pymupdf>=1.24.0
python-docx>=0.8.11
reportlab>=4.1.0
tiktoken>=0.6.0
python-dotenv>=1.0.0
plotly>=5.20.0
pandas>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
